Do not mark already reviewed file as reviewed

+++ polygerrit-ui/app/elements/diff/gr-diff-view/gr-diff-view.ts
@@ -462,6 +462,7 @@ export class GrDiffView extends KeyboardShortcutMixin(PolymerElement) {
     this.$.reviewed.checked = reviewed;
     if (!this._patchRange?.patchNum || !this._path) return;
     const path = this._path;
+    if (this._reviewedFiles.has(path) && reviewed) return;
     if (reviewed) this._reviewedFiles.add(path);
     else this._reviewedFiles.delete(path);
     this._saveReviewedState(reviewed).catch(err => {
